# Backend - nginx/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Реализация RestAPI на Go с использование Swagger для документации",
        "title": "School 21 ApiFromSpec docs",
        "termsOfService": "http://swagger.io/terms",
        "contact": {
            "name": "desmonsu"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/clients": {
            "get": {
                "description": "Возвращает массив клиентов с метаинформацией о пагинации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Получает всех клиентов из базы данных с пагинацией",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "список клиентов с метаинформацией",
                        "schema": {
                            "$ref": "#/definitions/apifromspec_internal_pkg.GetClientsMeta"
                        }
                    },
                    "500": {
                        "description": "failed to fetch clients",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Возвращает информацию об успешности операции",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Добавление клиента в базу данных",
                "parameters": [
                    {
                        "description": "Данные для добавления клиента",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apifromspec_internal_pkg.AddClientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "client created successfully",
                        "schema": {
                            "$ref": "#/definitions/apifromspec_internal_pkg.GetClientsResponse"
                        }
                    },
                    "400": {
                        "description": "wrong input format or validation error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "client already exists",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to add client",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/clients/search": {
            "get": {
                "description": "Возвращает массив клиентов, соответствующих запросу. Если клиенты не найдены, возвращает пустой массив.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Получает клиентов из базы данных по имени и фамилии",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя клиента",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Фамилия клиента",
                        "name": "surname",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "список клиентов (может быть пустым)",
                        "schema": {
                            "$ref": "#/definitions/apifromspec_internal_pkg.GetClientsResponse"
                        }
                    },
                    "400": {
                        "description": "name and surname are required",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to fetch client",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/clients/{id}": {
            "delete": {
                "description": "Возвращает информацию об успешности операции",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Удаляет клиента из базы данных",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID клиента для удаления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "client deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "wrong input format or can not parse id from URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "client not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to delete client",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/clients/{id}/address": {
            "patch": {
                "description": "Заменяет адрес клиента в базе данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Смена адреса для клиента",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID клиента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Параметры нового адреса клиента",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apifromspec_internal_pkg.AddressChangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "address changed successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "wrong input format or validation error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "id not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to process address",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apifromspec_internal_domain.Gender": {
            "type": "string",
            "enum": [
                "M",
                "F"
            ],
            "x-enum-varnames": [
                "Male",
                "Female"
            ]
        },
...
        "apifromspec_internal_pkg.AddClientRequest": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string",
                    "example": "1990-01-15T00:00:00Z"
                },
                "city": {
                    "type": "string",
                    "example": "Moscow"
                },
                "client_name": {
                    "type": "string",
                    "example": "Ivan"
                },
                "client_surname": {
                    "type": "string",
                    "example": "Petrov"
                },
                "country": {
                    "type": "string",
                    "example": "Russia"
                },
                "gender": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/apifromspec_internal_domain.Gender"
                        }
                    ],
                    "example": "M"
                },
                "street": {
                    "type": "string",
                    "example": "Tverskaya"
                }
            }
        },
...
        "apifromspec_internal_pkg.AddressChangeRequest": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Saint Petersburg"
                },
                "country": {
                    "type": "string",
                    "example": "Russia"
                },
                "street": {
                    "type": "string",
                    "example": "Nevskiy Prospect"
                }
            }
        },
...
        "apifromspec_internal_pkg.GetClientsMeta": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apifromspec_internal_pkg.GetClientsResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/apifromspec_internal_pkg.Pagination"
                }
            }
        },
        "apifromspec_internal_pkg.GetClientsResponse": {
            "type": "object",
            "properties": {
                "address_id": {
                    "type": "string",
                    "example": "660f8400-e29b-41d4-a716-446655440011"
                },
                "birthday": {
                    "type": "string",
                    "example": "1990-01-15T00:00:00Z"
                },
                "client_name": {
                    "type": "string",
                    "example": "Ivan"
                },
                "client_surname": {
                    "type": "string",
                    "example": "Petrov"
                },
                "gender": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/apifromspec_internal_domain.Gender"
                        }
                    ],
                    "example": "M"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "registration_date": {
                    "type": "string",
                    "example": "2026-04-28T10:00:00Z"
                }
            }
        },
...
        "apifromspec_internal_pkg.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 3
                },
                "next_offset": {
                    "type": "integer",
                    "example": 1
                },
                "offset": {
                    "type": "integer",
                    "example": 0
                },
                "total": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
    }
}